debian/*: adding package guh-tests
This commit readd in addition to e02d8ad09ef8a34b7f823cc51fadaa632454abca
the package guh-tests to reflect the the various testing binarys as a
additional package.
The test will be found in /usr/lib/guh/tests as the binarys are
architecture dependent. The binarys are strippend on the RPATH value.