Re-enable the default ctor for BrowseResults
This opens up the possibility for the plugin dev to mess up the
request id if [s]he creates a new one instead of using the one given by
the api, but not being able to keep results in a Qt container
is not a price I am willing to pay for this. After all it'S going to be
very visible on the first test to the plugin dev if this deosn't match.