Update discovery API.
This streamlines the API with the Browsing API and other upcoming
api changes. Also it enables to do syncronous discoveries and
allows better tracking of discovery transactions. Last but not
least it simplifies the devicesDiscovered() and autoDevicesAppeared()
signals by not requiring the duplicate DeviceClassId parameter any
more and with that, allows to pop up multiple devices with different
device classes at once.
Also this would allow to include e.g. Philips hue lights in the
"Add things" wizard at the discovery as such a discovery now
can return hue bridges too.