Add rpath to nymead binary
This allows to run it from the build directory without having to
specify LD_LIBRARY_PATH.

This is particularly useul when working with parts that require
special capabilities such as the NetworkDiscovery, given that
setcap on a binary disables LD_LIBRARY_LATH loading for security
reasons.

Building the dpkg package, the rpath will not be added.